requirements.txt
fastapi
uvicorn
tensorflow
numpy
Pillow
flask
gunicorn